backend-core/auth-service/requirements.txt
# Flask y extensiones
Flask==3.0.0
Flask-CORS==6.0.0
Flask-PyMongo==2.3.0
Flask-Limiter==3.5.0
Flask-RESTX==1.3.0

# Base de datos
pymongo==4.6.1

# Autenticación y seguridad
PyJWT==2.8.0
bcrypt==4.1.2
google-auth==2.23.3 # Added for Google Sign-In

# Validaciones
email-validator==2.1.0

# Email
secure-smtplib==0.1.1

# Variables de entorno
python-dotenv==1.0.0

# File handling
Pillow==10.2.0
werkzeug==3.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-asyncio==0.21.1

# Desarrollo
python-json-logger==2.0.7
